pyproject.toml
[tool.poetry]
name = "critical_projects"
version = "0.1.0"
description = "A tool to access criticality of code dependencies based on page rank in a dependency graph"
authors = ["HelgeCPH <ropf@itu.dk>"]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^5.3.1"
psutil = "^5.7.3"
neo4j = "^4.2.0"
criticality-score = "^1.0.4"
gsutil = "^4.57"
truckfactor = "^0.2.0"
pandas = "^1.1.5"

[tool.poetry.dev-dependencies]
ipython = "^7.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
